+++ rpc/client/local/local.go
@@ -2,6 +2,7 @@ package local
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	cmtpubsub "github.com/cometbft/cometbft/internal/pubsub"
@@ -9,7 +10,6 @@ import (
 	"github.com/cometbft/cometbft/libs/bytes"
 	"github.com/cometbft/cometbft/libs/log"
 	nm "github.com/cometbft/cometbft/node"
-	"github.com/cometbft/cometbft/rpc"
 	rpcclient "github.com/cometbft/cometbft/rpc/client"
 	"github.com/cometbft/cometbft/rpc/core"
 	ctypes "github.com/cometbft/cometbft/rpc/core/types"
@@ -59,6 +59,18 @@ func New(node *nm.Node) *Local {
 }
 
 var _ rpcclient.Client = (*Local)(nil)
+
+type ErrParseQuery struct {
+	Source error
+}
+
+func (e ErrParseQuery) Error() string {
+	return fmt.Sprintf("failed to parse query: %v", e.Source)
+}
+
+func (e ErrParseQuery) Unwrap() error {
+	return e.Source
+}
 
 // SetLogger allows to set a logger on the client.
 func (c *Local) SetLogger(l log.Logger) {
@@ -220,7 +232,7 @@ func (c *Local) Subscribe(
 ) (out <-chan ctypes.ResultEvent, err error) {
 	q, err := cmtquery.New(query)
 	if err != nil {
-		return nil, rpc.ErrParseQuery{Source: err}
+		return nil, ErrParseQuery{Source: err}
 	}
 
 	outCap := 1
@@ -300,7 +312,7 @@ func (c *Local) resubscribe(subscriber string, q cmtpubsub.Query) types.Subscrip
 func (c *Local) Unsubscribe(ctx context.Context, subscriber, query string) error {
 	q, err := cmtquery.New(query)
 	if err != nil {
-		return rpc.ErrParseQuery{Source: err}
+		return ErrParseQuery{Source: err}
 	}
 	return c.EventBus.Unsubscribe(ctx, subscriber, q)
 }
