+++ rpc/jsonrpc/server/http_server.go
@@ -17,9 +17,16 @@ import (
 	"golang.org/x/net/netutil"
 
 	"github.com/cometbft/cometbft/libs/log"
-	jsonrpcerrors "github.com/cometbft/cometbft/rpc/jsonrpc/errors"
 	types "github.com/cometbft/cometbft/rpc/jsonrpc/types"
 )
+
+type ErrInvalidRemoteAddress struct {
+	Addr string
+}
+
+func (e ErrInvalidRemoteAddress) Error() string {
+	return fmt.Sprintf("invalid listening address %s (use fully formed addresses, including the tcp:// or unix:// prefix)", e.Addr)
+}
 
 // Config is a RPC server configuration.
 type Config struct {
@@ -263,7 +270,7 @@ func (h maxBytesHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 func Listen(addr string, maxOpenConnections int) (listener net.Listener, err error) {
 	parts := strings.SplitN(addr, "://", 2)
 	if len(parts) != 2 {
-		return nil, jsonrpcerrors.ErrInvalidRemoteAddress{Addr: addr}
+		return nil, ErrInvalidRemoteAddress{Addr: addr}
 	}
 	proto, addr := parts[0], parts[1]
 	listener, err = net.Listen(proto, addr)
